GoKoala CRD support: Split url + language to own files
